backend/utils/exif_utils.py
from PIL import Image
from PIL.ExifTags import TAGS, GPSTAGS
from typing import Optional, Tuple

def get_gps_coordinates(image_path: str) -> Optional[Tuple[float, float]]:
    """Extract GPS coordinates from image EXIF data."""
    try:
        image = Image.open(image_path)
        exif_data = image._getexif()
        
        if not exif_data:
            return None
        
        gps_info = {}
        for tag, value in exif_data.items():
            tag_name = TAGS.get(tag, tag)
            if tag_name == "GPSInfo":
                for gps_tag in value:
                    gps_tag_name = GPSTAGS.get(gps_tag, gps_tag)
                    gps_info[gps_tag_name] = value[gps_tag]
        
        if not gps_info:
            return None
        
        lat = _convert_to_degrees(gps_info.get("GPSLatitude"))
        lng = _convert_to_degrees(gps_info.get("GPSLongitude"))
        
        if lat and lng:
            if gps_info.get("GPSLatitudeRef") == "S":
                lat = -lat
            if gps_info.get("GPSLongitudeRef") == "W":
                lng = -lng
            return (lat, lng)
        
        return None
    except Exception as e:
        print(f"Error extracting GPS: {e}")
        return None

def _convert_to_degrees(value):
    """Convert GPS coordinates to degrees."""
    if not value:
        return None
    d, m, s = value
    return float(d) + float(m) / 60.0 + float(s) / 3600.0
