backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
google-generativeai==0.3.2
qdrant-client==1.12.1
firebase-admin==6.6.0
supabase==2.10.0
pillow==11.0.0
python-dotenv==1.0.1
pydantic==2.10.3
httpx==0.27.2
PyPDF2==3.0.1
python-docx==1.1.0
python-socketio==5.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
